+++ src/assets/Pages/CrearPlaylist/CrearPlaylist.css
@@ -1,0 +1,135 @@
+.crear-playlist-container {
+    width: 300px;
+    height: 260px;
+    display: flex;
+    flex-direction: column;
+}
+
+.cpc-tituloA {
+    text-align: center;
+    padding-top: 10px;
+}
+
+.cpc-subtituloB {
+    text-align: start;
+    padding-top: 10px;
+}
+
+.crear-playlist-container h3 {
+    font-size: 26px;
+    font-weight: 500;
+    line-height: 36px;
+    letter-spacing: 0.5px;
+    margin: 0;
+}
+
+.crear-playlist-container h5 {
+    font-size: 20px;
+    font-weight: 500;
+    line-height: 36px;
+    letter-spacing: 0.5px;
+    margin: 0;
+}
+
+.cpc-input-playlist {
+    width: 100%;
+    height: auto;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-top: 10px;
+}
+
+.cpc-input-playlist input {
+    border: 1px solid;
+    width: 100%;
+    height: 48px;
+    padding: 0px 16px 0px 16px;
+    border-radius: 0px 16px 0px 16px;
+    border-color: #26262E;
+}
+
+.cpc-input-playlist input:focus {
+    border: 2px solid #FF8E0A; /* Cambiamos el color del borde a anaranjado */
+    width: 100%;
+    height: 48px;
+    padding: 0px 16px;
+    border-radius: 0px 16px 0px 16px;
+    outline: none;
+}
+
+.cpc-crearplaylist-footer{
+    width: 100%;
+    height: auto;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-top: 20%;
+}
+
+.fade-in-right {
+    -webkit-animation: fade-in-right 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) both;
+    animation: fade-in-right 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) both;
+  }
+  
+  @-webkit-keyframes fade-in-right {
+    0% {
+      -webkit-transform: translateX(50px);
+      transform: translateX(50px);
+      opacity: 0;
+    }
+  
+    100% {
+      -webkit-transform: translateX(0);
+      transform: translateX(0);
+      opacity: 1;
+    }
+  }
+  
+  @keyframes fade-in-right {
+    0% {
+      -webkit-transform: translateX(50px);
+      transform: translateX(50px);
+      opacity: 0;
+    }
+  
+    100% {
+      -webkit-transform: translateX(0);
+      transform: translateX(0);
+      opacity: 1;
+    }
+  }
+  
+  .fade-in-left {
+    -webkit-animation: fade-in-left 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) both;
+    animation: fade-in-left 0.6s cubic-bezier(0.39, 0.575, 0.565, 1) both;
+  }
+  
+  @-webkit-keyframes fade-in-left {
+    0% {
+      -webkit-transform: translateX(-50px);
+      transform: translateX(-50px);
+      opacity: 0;
+    }
+  
+    100% {
+      -webkit-transform: translateX(0);
+      transform: translateX(0);
+      opacity: 1;
+    }
+  }
+  
+  @keyframes fade-in-left {
+    0% {
+      -webkit-transform: translateX(-50px);
+      transform: translateX(-50px);
+      opacity: 0;
+    }
+  
+    100% {
+      -webkit-transform: translateX(0);
+      transform: translateX(0);
+      opacity: 1;
+    }
+  }
+
