+++ src/assets/components/BotonesDani/BotonesDaniGris.css
@@ -1,0 +1,36 @@
+.botonesDani-btn-grey {
+    background-color: #E4E6E8;
+    width: 320px;
+    height: 48px;
+    padding: 12px, 16px, 12px, 16px;
+    border-radius: 0px 16px 0px 16px;
+    gap: 12px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-family: Rubik;
+    margin-top: 15px;
+}
+
+.botonesDani-btn-grey a{
+color: aliceblue;
+font-size: 16px;
+font-weight: 500;
+line-height: 24px;
+letter-spacing: 0.5px;
+text-align: center;
+}
+
+.botonesDani-btn-grey:hover {
+    background-color: #FF8E0A;
+    width: 320px;
+    height: 48px;
+    padding: 12px, 16px, 12px, 16px;
+    border-radius: 0px 16px 0px 16px;
+    gap: 12px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-family: Rubik;
+    margin-top: 15px;
+}
