+++ src/assets/components/BotonesDani/BotonesDaniGris.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import './BotonesDaniGris.css';
+
+const BotonesDaniGris = (props) => {
+  return (
+    <>
+      <div className="botonesDani-btn-grey">
+        <a href="">{props.opcion}</a>
+      </div>
+    </>
+  )
+}
+
+export default BotonesDaniGris
