+++ src/assets/components/BotonesDani/BotonesDaniNaranja.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import './BotonesDaniNaranja.css';
+
+const BotonesDaniNaranja = (props) => {
+  return (
+    <>
+      <div className="botonesDani-btn-organge">
+        <a href="">{props.opcion}</a>
+      </div>
+    </>
+  )
+}
+
+export default BotonesDaniNaranja
