+++ src/assets/components/BotonesDani/BotonesDaniNegro.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import './BotonesDaniNegro.css';
+
+const BotonesDaniNegro = (props) => {
+  return (
+    <>
+      <div className="botonesDani-btn-black">
+        <a href="">{props.opcion}</a>
+      </div>
+    </>
+  )
+}
+
+export default BotonesDaniNegro
